fix: javaDoc for sqrt changed, param and return types in sqrt to double, mem usage in calcWithMem changed by one additional method, test for calcWithMem is rewrote

+++ homework/src/homework3/Calculator.java
@@ -72,10 +72,10 @@ public interface Calculator {
     double absoluteValue(double value);
     
     /**
-     * Нахождение квадратного корня из целого числа.
-     * @param value целое неотрицательное число
-     * @return целое число, квадрат которого
+     * Нахождение квадратного корня из числа.
+     * @param value неотрицательное число
+     * @return число, квадрат которого
      * меньше или равен числу-параметру.
      */
-    int squareRoot(int value);
+    double squareRoot(double value);
 }
